[Task 73139] added equals and hashcode methods to IndexableObjects
